logger removed and moved request param binder to core for other services to use filter with encoded string
