+++ libs/core/src/main/java/com/sourcefuse/jarc/core/filters/webdatabinder/RequestParamsBinder.java
@@ -2,22 +2,23 @@ package com.sourcefuse.jarc.services.usertenantservice.core;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.beans.PropertyEditorSupport;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 
+import java.beans.PropertyEditorSupport;
+
 @Slf4j
-public class GenericEditor<T> extends PropertyEditorSupport {
+public class RequestParamsBinder<T> extends PropertyEditorSupport {
 
   private ObjectMapper objectMapper;
   private Class<T> targetType;
 
-  public GenericEditor(ObjectMapper objectMapper, Class<T> targetType) {
+  public RequestParamsBinder(ObjectMapper objectMapper, Class<T> targetType) {
     this.objectMapper = objectMapper;
     this.targetType = targetType;
   }
 
-  public GenericEditor(ObjectMapper objectMapper) {
+  public RequestParamsBinder(ObjectMapper objectMapper) {
     this.objectMapper = objectMapper;
   }
 
@@ -28,7 +29,7 @@ public class GenericEditor<T> extends PropertyEditorSupport {
     } else {
       T obj;
       try {
-        log.info("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!" + text);
+        log.info("Request params received in request" + text);
         obj = objectMapper.readValue(text, targetType);
       } catch (JsonProcessingException e) {
         throw new IllegalArgumentException(e);
