+++ services/user-tenant-service/src/test/java/com/sourcefuse/jarc/services/usertenantservice/integration/RoleUserTenantControllerTests.java
@@ -11,11 +11,7 @@ import com.sourcefuse.jarc.services.usertenantservice.mocks.MockRole;
 import com.sourcefuse.jarc.services.usertenantservice.mocks.MockTenantUser;
 import com.sourcefuse.jarc.services.usertenantservice.repository.RoleRepository;
 import com.sourcefuse.jarc.services.usertenantservice.repository.RoleUserTenantRepository;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import com.sourcefuse.jarc.services.usertenantservice.service.UserTenantService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -36,6 +32,12 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
 @DisplayName("Create Role User Tenants Apis Integration/units Tests")
 @ExtendWith(MockitoExtension.class)
 class RoleUserTenantControllerTests {
@@ -49,6 +51,9 @@ class RoleUserTenantControllerTests {
   @Mock
   private QueryService queryService;
 
+  @Mock
+  private UserTenantService userTenantService;
+
   @InjectMocks
   private RoleUserTenantController roleUserTenantController;
 
@@ -58,6 +63,7 @@ class RoleUserTenantControllerTests {
   private MockMvc mockMvc;
 
   private String basePath = "/roles/{id}/user-tenants";
+  private Specification mockSpecificationTwo;
 
   @BeforeEach
   public void setup() {
@@ -73,6 +79,7 @@ class RoleUserTenantControllerTests {
     mockUserTenant.setId(MockTenantUser.USER_TENANT_ID);
     //Set Current LoggedIn User
     MockCurrentUserSession.setCurrentLoggedInUser(null, null, null);
+    Specification mockSpecificationTwo = null;
   }
 
   @Test
@@ -245,8 +252,6 @@ class RoleUserTenantControllerTests {
       new UserTenant(),
       new UserTenant()
     );
-    Specification mockSpecificationTwo =
-      MockCurrentUserSession.getSpecification(queryService);
     Mockito
       .when(roleUserTenantRepository.findAll(mockSpecificationTwo))
       .thenReturn(expectedTenantList);
@@ -271,8 +276,7 @@ class RoleUserTenantControllerTests {
   void testGetAllUserTenantByRole_Empty() throws Exception {
     // Arrange
     List<UserTenant> expectedTenantList = Arrays.asList();
-    Specification mockSpecificationTwo =
-      MockCurrentUserSession.getSpecification(queryService);
+
     Mockito
       .when(roleUserTenantRepository.findAll(mockSpecificationTwo))
       .thenReturn(expectedTenantList);
@@ -296,8 +300,8 @@ class RoleUserTenantControllerTests {
   @DisplayName("Test: case for count success")
   void testCount_Success() throws Exception {
     // Mock the behavior of roleRepository.count()
-    Specification mockSpecificationTwo =
-      MockCurrentUserSession.getSpecification(queryService);
+    Specification mockSpecificationTwo = null;
+    //  MockSpecification.getSpecification(queryService);
     Mockito
       .when(roleUserTenantRepository.findAll(mockSpecificationTwo))
       .thenReturn(Arrays.asList(new UserTenant(), new UserTenant())); // Mock a count of 2
@@ -321,8 +325,6 @@ class RoleUserTenantControllerTests {
   @DisplayName("Test case should pass for 0 count")
   void testCount_Empty() throws Exception {
     // Mock the behavior of roleRepository.count()
-    Specification mockSpecificationTwo =
-      MockCurrentUserSession.getSpecification(queryService);
     Mockito
       .when(roleUserTenantRepository.findAll(mockSpecificationTwo))
       .thenReturn(Arrays.asList()); // Mock a count of 0
@@ -354,8 +356,6 @@ class RoleUserTenantControllerTests {
       new UserTenant()
     );
 
-    Specification mockSpecificationTwo =
-      MockCurrentUserSession.getSpecification(queryService);
     Mockito
       .when(roleUserTenantRepository.findAll(mockSpecificationTwo))
       .thenReturn(userTenantArrayList);
@@ -392,8 +392,7 @@ class RoleUserTenantControllerTests {
     // Arrange
     UUID nonExistingGroupId = mockRoleId;
     UserTenant sourceUserTenant = this.mockUserTenant;
-    Specification mockSpecificationTwo =
-      MockCurrentUserSession.getSpecification(queryService);
+
     Mockito
       .when(roleUserTenantRepository.findAll(mockSpecificationTwo))
       .thenReturn(new ArrayList());
@@ -420,9 +419,6 @@ class RoleUserTenantControllerTests {
   @Test
   @DisplayName("Test:Delete existing Role UserTenant by Id")
   void testDeleteRoleById_Success() throws Exception {
-    Specification mockSpecificationTwo =
-      MockCurrentUserSession.getSpecification(queryService);
-
     Mockito
       .when(roleUserTenantRepository.delete(mockSpecificationTwo))
       .thenReturn(2L);
@@ -444,8 +440,6 @@ class RoleUserTenantControllerTests {
   @DisplayName("Delete non-existing roles UserTenant should return 404 status")
   void testDeleteRolesById_NotFound() throws Exception {
     // Mock the behavior of dependencies
-    Specification mockSpecificationTwo =
-      MockCurrentUserSession.getSpecification(queryService);
 
     Mockito
       .doThrow(new ResponseStatusException(HttpStatus.NOT_FOUND))
