+++ services/user-tenant-service/src/test/java/com/sourcefuse/jarc/services/usertenantservice/integration/UserTenantPrefsControllerTests.java
@@ -1,27 +1,14 @@
 package com.sourcefuse.jarc.services.usertenantservice.integration;
 
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
 import com.sourcefuse.jarc.core.filters.services.QueryService;
 import com.sourcefuse.jarc.services.usertenantservice.controller.UserTenantPrefsController;
 import com.sourcefuse.jarc.services.usertenantservice.dto.UserTenantPrefs;
 import com.sourcefuse.jarc.services.usertenantservice.mocks.JsonUtils;
 import com.sourcefuse.jarc.services.usertenantservice.mocks.MockCurrentUserSession;
+import com.sourcefuse.jarc.services.usertenantservice.mocks.MockSpecification;
 import com.sourcefuse.jarc.services.usertenantservice.mocks.MockTenantUser;
 import com.sourcefuse.jarc.services.usertenantservice.repository.UserTenantPrefsRepository;
 import com.sourcefuse.jarc.services.usertenantservice.service.UserTenantPrefsService;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -34,6 +21,21 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.UUID;
+
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @DisplayName("User Tenant Prefs Controller Integration Tests")
 @ExtendWith(MockitoExtension.class)
@@ -112,7 +114,7 @@ class UserTenantPrefsControllerTests {
       userTenantPrefs1,
       userTenantPrefs2
     );
-    Specification mockSpecification = MockCurrentUserSession.getSpecification(
+    Specification mockSpecification = MockSpecification.getSpecification(
       queryService
     );
     // Mock the repository
@@ -140,7 +142,7 @@ class UserTenantPrefsControllerTests {
   void testGetAllUsTenantPrefs_Empty() throws Exception {
     // Arrange
     List<UserTenantPrefs> userTenantPrefsList = new ArrayList<>();
-    Specification mockSpecification = MockCurrentUserSession.getSpecification(
+    Specification mockSpecification = MockSpecification.getSpecification(
       queryService
     );
     // Mock the repository
